remove old style extension support
