+++ middleman-core/lib/middleman-more/core_extensions/default_helpers.rb
@@ -248,3 +248,5 @@ class Middleman::CoreExtensions::DefaultHelpers < ::Middleman::Extension
     end
   end
 end
+
+Middleman::CoreExtensions::DefaultHelpers.register
